+++ src/index.ts
@@ -2,5 +2,6 @@ export * from './core.js';
 export * from './server.js';
 export { createExpressGraphQLMiddleware } from './express.js';
 export { createFetchGraphQLHandler } from './fetch.js';
+export * from './hasura.js';
 export { createHonoGraphQLHandler } from './hono.js';
 export { createNextGraphQLHandler } from './next.js';